requirements.txt
addict==2.4.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
auditwheel==5.1.2
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.6
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.7.0
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
ez-setup==0.9
fonttools==4.30.0
hausdorff==0.2.6
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.11.1
jupyter-server==1.15.4
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
lazy-object-proxy==1.7.1
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclassic==0.3.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.10
notebook-shim==0.1.0
numba==0.56.2
numpy==1.21.6
open3d==0.15.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
pptk==0.1.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyelftools==0.28
pygame==2.1.2
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.7
PyQt5-Qt5==5.15.2
pyquaternion==0.9.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2020.1
pywin32==305
pywinpty==2.0.9
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
QtPy==1.9.0
requests==2.27.1
rope==0.23.0
Rtree==0.9.7
ruamel.yaml==0.15.87
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
tbb==2021.5.1
terminado==0.13.3
testpath==0.4.4
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.10.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typed-ast==1.5.4
typing_extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.13.3
xlwt==1.3.0
yapf==0.32.0
yarg==0.1.9
zipp==3.7.0
zmq==0.0.0
DracoPy==1.2.0